feat: better layout, new styles and file renaming
