+++ src/components/javigaralva/password-generator-app.css
@@ -1,0 +1,3 @@
+.javigaralva-no-tap-highlight {
+	-webkit-tap-highlight-color: transparent;
+}
