+++ src/components/javigaralva/utils/is-number.tsx
@@ -1,4 +1,4 @@
-import { NUMBERS } from '../constants'
+import { NUMBERS } from '../constants/constants'
 
 export function isNumber(letter: string) {
 	return NUMBERS.includes(letter)
