+++ src/components/javigaralva/utils/is-symbol.tsx
@@ -1,4 +1,4 @@
-import { SYMBOLS } from '../constants'
+import { SYMBOLS } from '../constants/constants'
 
 export function isSymbol(letter: string) {
 	return SYMBOLS.includes(letter)
