+++ src/components/javigaralva/utils/is-upper-case-letter.tsx
@@ -1,4 +1,4 @@
-import { UPPER_CASE_LETTERS_LIST } from '../constants'
+import { UPPER_CASE_LETTERS_LIST } from '../constants/constants'
 
 export function isUpperCaseLetter(letter: string) {
 	return UPPER_CASE_LETTERS_LIST.includes(letter)
