+++ src/components/javigaralva/utils/style-generators/big-circle-doted.tsx
@@ -1,12 +1,12 @@
 import { CSSProperties } from 'react'
-import { random } from '../utils/random'
+import { random } from '../random'
 
-export function getBigDotsStyle() {
+export function getBigCircleDottedStyle() {
 	const style: CSSProperties = {}
-	const width = 128 + random(-100, 100)
-	const offset = 154 + random(-100, 100)
-	const alpha = 2 + random(-1, 6)
-	const color = 0 + random(-10, 40)
+	const width = 8 + random(-4, 2)
+	const offset = 204 + random(-100, 100)
+	const alpha = 6 + random(-2, 2)
+	const color = 124 + random(-25, 75)
 	style.outlineOffset = `${offset}px`
 	style.outlineColor = `hsl(${color}deg 100% 50% / ${alpha}%)`
 	style.outlineWidth = `${width}px`
