+++ src/components/javigaralva/utils/style-generators/big-dots.tsx
@@ -1,12 +1,12 @@
 import { CSSProperties } from 'react'
-import { random } from '../utils/random'
+import { random } from '../random'
 
-export function getBigCircleDottedStyle() {
+export function getBigDotsStyle() {
 	const style: CSSProperties = {}
-	const width = 8 + random(-4, 2)
-	const offset = 204 + random(-100, 100)
-	const alpha = 6 + random(-2, 2)
-	const color = 124 + random(-25, 75)
+	const width = 128 + random(-100, 100)
+	const offset = 154 + random(-100, 100)
+	const alpha = 2 + random(-1, 6)
+	const color = 0 + random(-10, 40)
 	style.outlineOffset = `${offset}px`
 	style.outlineColor = `hsl(${color}deg 100% 50% / ${alpha}%)`
 	style.outlineWidth = `${width}px`
