+++ src/components/javigaralva/utils/style-generators/square-and-vertical-lines.tsx
@@ -1,5 +1,5 @@
 import { CSSProperties } from 'react'
-import { random } from '../utils/random'
+import { random } from '../random'
 
 export function getSquareAndVerticalLinesStyle() {
 	const style: CSSProperties = {}
